requirements.txt
numpy>=1.16.2
tqdm>=4.31.1
torchvision>=0.2.2
torchnet>=0.0.4
visdom>=0.1.8.8
fire>=0.1.3
opencv_python>=4.1.0.25
ipdb>=0.12
torch>=1.0.1
Pillow>=6.0.0
